moving client code to api server area
